Job class added
The job is a compositum of multiple commands. It contains a list
of commands and can execute them in order of adding. When doing an
undo, the undo() methods of the contained commands are performed
in reversed order.

The commands are supposed to be used for the actual player actions
and will be saved in the undo / redo stacks.

Example:
Moving left is a job which consists of several move and merge
commands as well as creating the new tile and incrementing points
counter afterwards. It is indivisible - the user shall be able to
undo the whole move, but not the particular tile movement / merging
commands.