Commands are now executed immediately
In order to avoid merges with tiles before moving them on a target
position and creating new tiles on non-free positions, the commands
are executed before enqueueing in the job. Therefore it is not
necessery to use second "live" board for tiles movement. When the
Job has to be undone or redone, the commands are already OK and
they can be executed at once as so before.