Glade files and themes moved to resources directory
The GUI definition file (Glade) and CSS style sheet moved to the separate
resources directory. The style sheets are now stored in separate direcrories
which can be used for theme selection. As the selection logic in not yet
implemented, the active theme "classic" is currently hard coded as a
function argument.