Score counter added to the game engine
Game engine has a new score counter variable now.

When two tiles merge, the value of the new (merged) tile
is added to the score counter in the game engine, just like
in the orginal 2048 game.

Score is added using a new AddScore command and is subtracted when
undoing the move