Sidebar removed and its content moved to the header bar
The header bar is more appropriate for the score and time than the
sidebar. Additionaly the sidebar distracts the user and take
unnecessarily some space of the main game window. Therefore it is now deleted and the game score and time is moved to the headerbar of the main window.

The old sidebar controller, view and style definition is removed from the project and the translation files updated.

Closes #25 Remove the sidebar